+++ packages/node-client/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @stoker-platform/node-client
 
+## 0.5.79
+
+### Patch Changes
+
+- Updated dependencies
+    - @stoker-platform/types@0.5.58
+    - @stoker-platform/utils@0.5.70
+
 ## 0.5.78
 
 ### Patch Changes