+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @stoker-platform/utils
 
+## 0.5.70
+
+### Patch Changes
+
+- Updated dependencies
+    - @stoker-platform/types@0.5.58
+
 ## 0.5.69
 
 ### Patch Changes